pyproject.toml
[tool.poetry]
name = "haus-utils"
version = "0.1.0"
description = "HAUS Utilities"
authors = ["iTecAI <matteovh@gmail.com>"]
license = "GPL v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beanie = "^1.24.0"
pydantic = "^2.5.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
